xccl/core: Bug fix in the OOB allgather code
The OOB allgather code expects each caller process to pass in its
rank, based on the API.  But the current implementation simply passes
"0" for all processes.  This worked fine so far because none of the
application codes used that value, but it could break a future code
that relies on that parameter.

Signed-off-by: Pavan Balaji <pavanbalaji.work@gmail.com>