xccl/mpod: Initial code draft for multi-POD collectives
This patch creates a new teamlib called "mpod" that essentially
provides a virtualization layer on top of UCX and NCCL.  For all user
objects, it returns virtual objects and internally maps these objects
to a combination of NCCL and UCX objects.  Similarly, for
user-provided OOB operations, MPOD provides a virtualization layer
that hides these functions from NCCL/UCX and provides virtual
functions to them.

The intent of this teamlib is to allow XCCL to be used in environments
where we want to use NCCL on more than one node (within a POD), and
UCX across PODs.

Signed-off-by: Pavan Balaji <pavanbalaji.work@gmail.com>